Fixed sounds not playing (out-of-range pan values)
map():
"As shown in the second example, numbers outside of the range are not clamped to the minimum and maximum parameters values, because out-of-range values are often intentional and useful."

went into debug mode - hand.scaledX sometimes ends up being -204 or so
(the tracking still works and current v. previous X changes consistently - but the values are negative??)
the only way i could be calculating negative scaledX values is if the kinect is giving *me* negative X values to begin with
no idea why that would be happening - but if i clamp the coordinates to their mins and maxs i should be fine?

(actually i... don't wanna break my code before night two so i'll just clamp the pan value for sound effects)